perf(core): cache getNavigator() result per router via WeakMap (#271)
